refactor(focus): extract FocusSessionStateMachine; centralize transitions; wrap unawaited calls

+++ lib/features/focus/domain/repositories/i_focus_session_repository.dart
@@ -4,7 +4,7 @@ abstract class IFocusSessionRepository {
   Future<FocusSession> startSession(FocusSession session);
   Future<void> updateSession(FocusSession session);
   Future<FocusSession?> getActiveSession();
-  Stream<List<FocusSession>> watchSessionsByTask(BigInt taskId);
+  Stream<List<FocusSession>> watchSessionsByTask(int taskId);
   Stream<List<FocusSession>> watchAllSessions();
-  Future<void> deleteSession(BigInt id);
+  Future<void> deleteSession(int id);
 }
